lagt en ny util som formatterer periode for et object med fom og tom
