+++ src/frontend/barnetilsyn/steg/5-pass-av-dine-barn/BarnepassSpørsmål.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Alert, DatePicker, Heading, HStack, Label, useDatepicker, VStack } from '@navikt/ds-react';
+import { Alert, Heading, Label, VStack } from '@navikt/ds-react';
 
 import BarnOver9År from './BarnOver9År';
 import {
@@ -11,14 +11,12 @@ import {
     errorKeyUtgifterTom,
 } from './passBarnVedleggUtils';
 import { BarnepassIntern } from './typer';
+import UtgifterDato from './utgifterDato';
 import LocaleRadioGroup from '../../../components/Teksthåndtering/LocaleRadioGroup';
 import LocaleTekst from '../../../components/Teksthåndtering/LocaleTekst';
 import { Barn, PassType } from '../../../typer/barn';
-import { EnumFelt } from '../../../typer/skjema';
-import { JaNei } from '../../../typer/søknad';
 import { Locale } from '../../../typer/tekst';
 import { Valideringsfeil } from '../../../typer/validering';
-import { nullableTilDato, tilLocaleDateString } from '../../../utils/formateringUtils';
 import { barnepassTekster } from '../../tekster/barnepass';
 
 interface Props {
@@ -38,44 +36,6 @@ const BarnepassSpørsmål: React.FC<Props> = ({
     nullstillValideringsfeil,
     locale,
 }) => {
-    const defaultHarUtgifterTilPass: EnumFelt<JaNei> = {
-        verdi: 'NEI',
-        label: 'Har utgifter til pass',
-        svarTekst: 'Nei',
-        alternativer: ['JA', 'NEI'],
-    };
-    const { datepickerProps: datepickerPropsFom, inputProps: inputPropsFom } = useDatepicker({
-        defaultSelected: nullableTilDato(barnepass.utgifter?.fom?.verdi),
-        onDateChange: (val) => {
-            const verdi = val ? { label: 'Fra', verdi: tilLocaleDateString(val) } : undefined;
-            oppdaterBarnMedBarnepass({
-                ...barnepass,
-                utgifter: {
-                    harUtgifterTilPass:
-                        barnepass.utgifter?.harUtgifterTilPass ?? defaultHarUtgifterTilPass,
-                    fom: verdi,
-                    tom: barnepass.utgifter?.tom,
-                },
-            });
-            nullstillValideringsfeil(errorKeyUtgifterFom(barn));
-        },
-    });
-    const { datepickerProps: datepickerPropsTom, inputProps: inputPropsTom } = useDatepicker({
-        defaultSelected: nullableTilDato(barnepass.utgifter?.tom?.verdi),
-        onDateChange: (val) => {
-            const verdi = val ? { label: 'Til', verdi: tilLocaleDateString(val) } : undefined;
-            oppdaterBarnMedBarnepass({
-                ...barnepass,
-                utgifter: {
-                    harUtgifterTilPass:
-                        barnepass.utgifter?.harUtgifterTilPass ?? defaultHarUtgifterTilPass,
-                    fom: barnepass.utgifter?.fom,
-                    tom: verdi,
-                },
-            });
-            nullstillValideringsfeil(errorKeyUtgifterTom(barn));
-        },
-    });
     return (
         <VStack gap={'6'}>
             <LocaleRadioGroup
@@ -117,24 +77,14 @@ const BarnepassSpørsmål: React.FC<Props> = ({
             {barnepass.utgifter?.harUtgifterTilPass?.verdi === 'NEI' && (
                 <>
                     <Label>{barnepassTekster.utgifter_dato.label[locale]}</Label>
-                    <HStack gap={'4'}>
-                        <DatePicker {...datepickerPropsFom}>
-                            <DatePicker.Input
-                                id={valideringsfeil[errorKeyUtgifterFom(barn)]?.id}
-                                label={barnepassTekster.utgifter_dato.fom[locale]}
-                                error={valideringsfeil[errorKeyUtgifterFom(barn)]?.melding}
-                                {...inputPropsFom}
-                            />
-                        </DatePicker>
-                        <DatePicker {...datepickerPropsTom}>
-                            <DatePicker.Input
-                                id={valideringsfeil[errorKeyUtgifterTom(barn)]?.id}
-                                label={barnepassTekster.utgifter_dato.tom[locale]}
-                                error={valideringsfeil[errorKeyUtgifterTom(barn)]?.melding}
-                                {...inputPropsTom}
-                            />
-                        </DatePicker>
-                    </HStack>
+                    <UtgifterDato
+                        barn={barn}
+                        barnepass={barnepass}
+                        oppdaterBarnMedBarnepass={oppdaterBarnMedBarnepass}
+                        valideringsfeil={valideringsfeil}
+                        nullstillValideringsfeil={nullstillValideringsfeil}
+                        locale={locale}
+                    />
                 </>
             )}
